apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Unchanted", email = "omkkdwivedi@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "argopy>=1.3.0",
    "fastapi>=0.116.1",
    "google-genai>=1.34.0",
    "python-dateutil>=2.9.0.post0",
    "uvicorn[standard]>=0.35.0",
]

[project.scripts]
backend = "backend:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
